+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 **NEW FEATURES**
 
-* a default coordiate can now be selected for for multiple coordinates dimensions.
+* a default coordiate can now be selected for multiple coordinates dimensions.
 
 **BUGS FIXED**
 
 * alignment along several dimensions (issue #248)
+
+* to() and ito() methods have been fixed to work correctly (issue #255) 
 
 ## Version 0.2.13